+++ sshd-sftp/src/main/java/org/apache/sshd/sftp/client/extensions/openssh/OpenSSHFsyncExtension.java
@@ -28,7 +28,7 @@ import org.apache.sshd.sftp.client.extensions.SftpClientExtension;
  * Implements the &quot;fsync@openssh.com&quot; extension
  *
  * @author <a href="mailto:dev@mina.apache.org">Apache MINA SSHD Project</a>
- * @see    <A HREF="https://github.com/openssh/openssh-portable/blob/master/PROTOCOL">OpenSSH - section 10</A>
+ * @see    <A HREF="https://github.com/openssh/openssh-portable/blob/master/PROTOCOL">OpenSSH - section 4.6</A>
  */
 public interface OpenSSHFsyncExtension extends SftpClientExtension {
     void fsync(Handle fileHandle) throws IOException;