+++ CHANGES.md
@@ -20,9 +20,25 @@
 
 ## Potential compatibility issues
 
+* A **new** SFTP configuration property has been introduced that limits the maximum amount of data that can be sent in a single *SSH_FXP_WRITE* packet - default=256KB
+
+```java
+    /**
+     * Force the use of a max. packet length for {@link AbstractSftpSubsystemHelper#doWrite(Buffer, int)} protection
+     * against malicious packets
+     */
+    public static final Property<Integer> MAX_WRITE_DATA_PACKET_LENGTH
+            = Property.integer("sftp-max-writedata-packet-length", 256 * 1024);
+```
+
+This might cause SFTP write failures for clients that might have sent larger buffers and they have been accepted so far. If this happens, simply increase
+this value (though the choice of 256KB should be compatible with the vast majority of clients).
+
 ## Minor code helpers
 
 ## Behavioral changes and enhancements
 
 * [SSHD-1231](https://issues.apache.org/jira/browse/SSHD-1231) Public key authentication: wrong signature algorithm used (ed25519 key with ssh-rsa signature)
+* [SSHD-1233](https://issues.apache.org/jira/browse/SSHD-1233) Added support for "limits@openssh.com" SFTP extension
+